Backend: Add missing controllers, routes and middleware files
