+++ backend/src/controllers/payment.controller.js
@@ -1,15 +1,13 @@
-import { Request, Response } from 'express';
-import { paymentService } from '../services/payment.service';
-import logger from '../utils/logger';
+import { paymentService } from '../services/payment.service.js';
 
 export class PaymentController {
-  async processPayment(req: Request, res: Response) {
+  async processPayment(req, res) {
     try {
       const { email } = req.body;
 
       const result = await paymentService.processPayment(email);
       
-      logger.info(`Payment processed successfully for: ${email}`);
+      console.log(`Payment processed successfully for: ${email}`);
       
       res.status(201).json({
         success: true,
@@ -19,9 +17,9 @@ export class PaymentController {
         },
       });
     } catch (error) {
-      logger.error('Payment controller error:', error);
+      console.error('Payment controller error:', error);
       
-      if (error instanceof Error && error.message === 'User already exists') {
+      if (error.message === 'User already exists') {
         return res.status(409).json({
           success: false,
           error: 'User already exists',
