--- backend/src/models/user.model.ts
@@ -1,56 +1,0 @@
-import mongoose from 'mongoose';
-import bcrypt from 'bcrypt';
-
-export type IUser = {
-  email: string;
-  password: string;
-  createdAt: Date;
-  lastLoginAt?: Date;
-}
-
-const userSchema = new mongoose.Schema<IUser>({
-  email: {
-    type: String,
-    required: true,
-    unique: true,
-    trim: true,
-    lowercase: true,
-  },
-  password: {
-    type: String,
-    required: true,
-  },
-  createdAt: {
-    type: Date,
-    default: Date.now,
-  },
-  lastLoginAt: {
-    type: Date,
-  },
-}, {
-  timestamps: true,
-  toJSON: {
-    transform: (_, ret) => {
-      delete ret.password;
-      return ret;
-    },
-  },
-});
-
-userSchema.pre('save', async function(next) {
-  if (!this.isModified('password')) return next();
-  
-  try {
-    const salt = await bcrypt.genSalt(10);
-    this.password = await bcrypt.hash(this.password, salt);
-    next();
-  } catch (error) {
-    next(error as Error);
-  }
-});
-
-userSchema.methods.comparePassword = async function(candidatePassword: string): Promise<boolean> {
-  return bcrypt.compare(candidatePassword, this.password);
-};
-
-export const User = mongoose.model<IUser>('User', userSchema);
