--- backend/src/utils/logger.ts
@@ -1,14 +1,0 @@
-import pino from 'pino';
-import env from '../config/environment';
-
-const logger = pino({
-  level: env.NODE_ENV === 'production' ? 'info' : 'debug',
-  transport: {
-    target: 'pino-pretty',
-    options: {
-      colorize: true,
-    },
-  },
-});
-
-export default logger;
